Refactor: extract practice logic into service layer
- Move question fetching, evaluation, Elo, and XP logic to practice_service
- Simplify tutorials routes to handle only request validation and response
